+++ CHANGELOG.md
@@ -4,7 +4,8 @@
 
 ## flack 1.3.0  not yet released
 
-- GET /executions?exid=20230310
+- GET /executions?exid=net.ntt.finance-
+- GET /executions?dexid=20230310
 - Add Flack.on_unit_created(unit) callback